Fix AT+CIPRXGET decoding
The sim7000 does not follow the specification in the datasheet for
responding to AT+CIPRXGET. The datasheet indicates that there should
only be one `+CIPRXGET:` line, however there can be more than one if
data is not ready yet. In this case the sim7000 will respond with a
potentially arbitrary number of lines containing `+CIPRXGET: 1` before
responding with the expected response and data.